Detect run sizes.

+++ 0x30/Pit.h
@@ -22,7 +22,7 @@ public:
     void ScrollOne();
 
     void Swap(size_t x, size_t y);
-    int& RunAt(size_t x, size_t y);
+    size_t& RunAt(size_t x, size_t y);
     int HeightAt(size_t x, size_t y) const;
     Tile TileAt(size_t x, size_t y) const;
     size_t PitIndex(size_t x, size_t y) const;
@@ -39,10 +39,10 @@ private:
     void CheckForHorizontalRuns(bool& foundRun);
 
     std::array<Tile, cols * rows> tiles_;
-    std::array<int, cols * rows> runs_;
+    std::array<size_t, cols * rows> runs_;
     std::array<int, cols * rows> heights_;
     size_t firstRow_ = 0;
     std::function<int(int, int)>& rnd_;
     bool impacted_;
-    int run_;
+    size_t run_;
 };
